chore(automation): convert to ES modules
