+++ features/step_definitions/support/dataTable2Object.ts
@@ -1,5 +1,5 @@
 import {DataTable} from '@cucumber/cucumber';
-import {toDateOrString} from './date-utils';
+import {toDateOrString} from './date-utils.js';
 
 function setPropertyValue(
   obj: any,
