+++ features/step_definitions/support/gebeurtenissen-vragen-api-helpers.ts
@@ -1,4 +1,4 @@
-import {logger} from './logger';
+import {logger} from './logger.js';
 
 export async function getOudsteOngelezenGebeurtenisVoorAfnemer(
   afnemerAanduiding: string,
