+++ tsconfig.json
@@ -1,13 +1,9 @@
 {
   "extends": "./node_modules/gts/tsconfig-google.json",
   "compilerOptions": {
-    "target": "ES2020",
-    "module": "commonjs",
+    "target": "es6",
+    "module": "nodenext",
     "strict": true,
-    "esModuleInterop": true,
     "rootDir": "."
-  },
-  "include": [
-    "features/**/*.ts"
-  ]
+  }
 }
